requirements-dev.txt
-r requirements.txt
pytest>=7.4
pytest-dotenv>=0.5
ruff>=0.5.0
black>=24.3
isort>=5.13
libcst>=1.1
pyyaml>=6.0
pre-commit>=3.5
